\name{non.coi-methods}
\docType{methods}
\alias{non.coi-methods}
\alias{non.coi,nonmem-method}
\title{ Output coi file}
\description{
  Methods for function \code{non.coi} in Package `PKreport' 
}

\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access coi content in the \code{file.coi} slot. }
}}
\keyword{methods}

