\name{non.cor-methods}
\docType{methods}
\alias{non.cor-methods}
\alias{non.cor,nonmem-method}
\title{ Output cor file}
\description{
  Methods for function \code{non.cor} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access cor content in the \code{file.cor} slot. }
}}
\keyword{methods}

