\name{non.cov-methods}
\docType{methods}
\alias{non.cov-methods}
\alias{non.cov,nonmem-method}
\title{ Output cov file}
\description{
  Methods for function \code{non.cov} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access cov content in the \code{file.cov} slot. }
}}
\keyword{methods}

