\name{non.lst.meth-methods}
\docType{methods}
\alias{non.lst.meth-methods}
\alias{non.lst.meth,nonmem-method}
\title{ output method in lst file}
\description{
  Methods for function \code{non.lst.meth} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access method section (\#meth tag) in the \code{file.lst} slot.}
}}
\keyword{methods}

