\name{non.lst.objs-methods}
\docType{methods}
\alias{non.lst.objs-methods}
\alias{non.lst.objs,nonmem-method}
\title{ Output objective function standard deviation in lst file}
\description{
  Methods for function \code{non.lst.objs} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access objective function standard deviation (\#OBJS tag) in the \code{file.lst} slot. }
}}
\keyword{methods}

