\name{non.lst.objt-methods}
\docType{methods}
\alias{non.lst.objt-methods}
\alias{non.lst.objt,nonmem-method}
\title{ Output objective function in lst file}
\description{
  Methods for function \code{non.lst.objt} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{  Access objective function (\#OBJT tag) in the \code{file.lst} slot.}
}}
\keyword{methods}

