\name{non.lst.objv-methods}
\docType{methods}
\alias{non.lst.objv-methods}
\alias{non.lst.objv,nonmem-method}
\title{ Output objective function value in lst file}
\description{
  Methods for function \code{non.objv} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access objective function value (\#OBJV tag) in the \code{file.lst} slot.}
}}
\keyword{methods}

