\name{non.lst.term-methods}
\docType{methods}
\alias{non.lst.term-methods}
\alias{non.lst.term,nonmem-method}
\title{ output analysis section in lst file }
\description{
  Methods for function \code{non.lst.term} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access analysis section (\#TERM, \#TERE tag) in the \code{file.lst} slot. }
}}
\keyword{methods}

