\name{non.phi-methods}
\docType{methods}
\alias{non.phi-methods}
\alias{non.phi,nonmem-method}
\title{ Output phi file }
\description{
  Methods for function \code{non.phi} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access phi content in the \code{file.coi} slot. }
}}
\keyword{methods}

