\name{non.select-methods}
\docType{methods}
\alias{non.select-methods}
\alias{non.select,nonmem-method}
\title{ Output selected lines in lst files }
\description{
  Methods for function \code{non.select} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ output the selected lines in lst files. }
}}
\keyword{methods}

