\name{non.tab-methods}
\docType{methods}
\alias{non.tab-methods}
\alias{non.tab,nonmem-method}
\title{ Output tab content in tab file}
\description{
  Methods for function \code{non.tab} in Package `PKreport' 
}
\section{Methods}{
\describe{

\item{object = "nonmem"}{ Access tab content in the \code{file.tab} slot. }
}}
\keyword{methods}

