\name{PKadjust}
\alias{PKadjust}

\title{ Update figures}
\description{
  This function is to update or modify certain figures with figure ID for particular purpose.
}
\usage{
PKadjust(figno, save=FALSE,...)
}

\arguments{
  \item{figno}{ the figure number generated from PKshow(). }
  \item{save}{ logical. To only display the modified figure, users need chose FALSE; to
              save in the management system, users need choose TRUE. }
  \item{...}{ All new figure configurations for this figure number.}
}

\details{ Please see vignettes for details.

}
\value{

}

\author{Xiaoyong Sun}

\examples{

}
\keyword{ methods}

