\name{PKclean}
\alias{PKclean}

\title{ Clean related archives}
\description{
  This function is to delete all archives (file directories and figures)
  and clean global variables in R environment.
}
\usage{
PKclean()
}


\details{   Please see vignettes for details.

}

\author{Xiaoyong Sun}

\examples{
# PKclean()
}
\keyword{ methods}

