\name{PKcode}
\alias{PKcode}

\title{ Generate R scripts}
\description{
  This function is to generate R scripts. To improve efficiency and help users to
  generate high-quality figures, users have option to modify related R scripts to
  meet their specific requirements. All generated R scripts match the order of figures
  generated in PKshow().
}
\usage{
PKcode(filename="PKcode.txt")
}

\arguments{
  \item{filename}{ the file name to store the R scripts.}
}

\details{ Please see vignettes for details.

}

\author{Xiaoyong Sun}


\keyword{ methods}

