\name{PKnum}
\alias{PKnum}

\title{ Output numerical value }
\description{
  This function is to convert data frame or matrix (the output from non.select) to number.
}
\usage{
PKnum(exp.data)
}

\arguments{
  \item{exp.data}{ a data frame or matrix selected from lst file. }
}

\details{ Please see vignettes for details.

}


\author{Xiaoyong Sun}

\examples{
# exp.data <- non.select(myclass, c(50:56))
# options(scipen=100)
# PKnum(exp.data)
# options(scipen=-100)
# PKnum(exp.data)
}
\keyword{ methods}

