\name{PKoutput}
\alias{PKoutput}

\title{ Output all results to folders and files }
\description{
  This function is to translate analysis from PKreport() to folders and files in the current working directory.
}
\usage{
PKoutput(nonmemObj, table.Rowv, table.Colv)
}

\arguments{
  \item{nonmemObj}{ a object of class NONMEM. }
  \item{table.Rowv}{ determine whether the row dendrogram should be reordered in the heatmap-like table. }
  \item{table.Colv}{ determine whether the column dendrogram should be reordered in the heatmap-like table. }
  
}

\details{ Please see vignettes for details.

}


\author{Xiaoyong Sun}

\examples{
# PKoutput(nonmemObj)
}
\keyword{ methods}

