\name{PKreport.1}
\alias{PKreport.1}

\title{ Exploratory data analysis }
\description{
  This function is for exploratory data analysis.
}
\usage{
PKreport.1(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}

\details{

}
\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.1(pdata)
}
\keyword{ methods}

