\name{PKreport.2}
\alias{PKreport.2}

\title{ Individual plots }
\description{
  This function is for individual plots.
}
\usage{
PKreport.2(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}


\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.2(pdata)
}
\keyword{ methods}

