\name{PKreport.3}
\alias{PKreport.3}

\title{ Goodness-of-fit plots }
\description{
  This function is for goodness-of-fit plots.
}
\usage{
PKreport.3(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}


\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.3(pdata)
}
\keyword{ methods}

