\name{PKreport.4}
\alias{PKreport.4}

\title{ Structural model diagnostics }
\description{
  This function is for structural model diagnostics.
}
\usage{
PKreport.4(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}

\details{

}
\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.4(pdata)
}
\keyword{ methods}

