\name{PKreport.5}
\alias{PKreport.5}

\title{ Residual model diagnostics }
\description{
  This function is for residual model diagnostics.
}
\usage{
PKreport.5(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}


\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.5(pdata)
}
\keyword{ methods}

