\name{PKreport.6}
\alias{PKreport.6}

\title{ Parameters diagnostics }
\description{
  This function is for parameters diagnostics.
}
\usage{
PKreport.6(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}

\details{

}
\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.6(pdata)
}
\keyword{ methods}

