\name{PKreport.7}
\alias{PKreport.7}

\title{ Covariate model diagnostics}
\description{
  This function is for covariate model diagnostics.
}
\usage{
PKreport.7(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}

\details{

}
\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.7(pdata)
}
\keyword{ methods}

