\name{PKreport.8}
\alias{PKreport.8}

\title{ Random effects diagnostics}
\description{
  This function is for random effects diagnostics.
}
\usage{
PKreport.8(pdata)
}

\arguments{
  \item{pdata}{ a data frame.}
  
}


\value{
Input is a data frame.
}

\author{Xiaoyong Sun}

\examples{
# PKreport.8(pdata)
}
\keyword{ methods}

