\name{PKshow}
\alias{PKshow}

\title{ Display results }
\description{
  This function is to display results from analysis.
}
\usage{
PKshow(nonmemObj, table.Rowv, table.Colv)
}

\arguments{
  \item{nonmemObj}{ a object of class NONMEM. }
  \item{table.Rowv}{ determine whether the row dendrogram should be reordered in the heatmap-like table. }
  \item{table.Colv}{ determine whether the column dendrogram should be reordered in the heatmap-like table. }
}

\details{ Please see vignettes for details.

}

\author{Xiaoyong Sun}

\examples{
# PKshow(nonmemObj, table.Rowv=TRUE, table.Rowv=TRUE)

# Only generate figure report
# PKshow() 
}
\keyword{ methods}

